# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Burrow JSON API
  description: |
    The hand-written canonical OpenAPI v3 description of the Burrow JSON API
    (unified across v0.2.0, v0.3.0, and v0.4.0).

    Coverage is locked by `TestOpenAPI_RouteCoverage` in
    `internal/api/openapi_test.go`: every route the chi mux answers must
    appear in this document or the build fails. Health probes (`/healthz`,
    `/readyz`) and the doc-serving endpoints (`/api/v1/openapi.{yaml,json}`)
    are excluded by convention.

    Schemas are intentionally minimal — the goal is shape coverage for SDK
    code-generators (TypeScript-fetch, Go), not exhaustive validation. Full
    schemas can be filled in incrementally without invalidating downstream
    SDKs (additive only).
  version: 0.4.0
  license:
    name: MIT
servers:
  - url: https://{host}
    variables:
      host:
        default: localhost:8080
        description: Burrow control-plane host.
...
components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: burrow_session
      description: |
        Session cookie issued by POST /api/v1/auth/login. State-changing
        cookie-authed requests MUST also send the burrow_csrf double-submit
        header (`X-CSRF-Token`).
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: bua_<token>
      description: |
        Automation bearer token (`Authorization: Bearer bua_...`). Bypasses
        CSRF (the token IS the CSRF defense). Permissions are the
        intersection of the user's current role and the set declared at
        mint time.
    CSRFToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: |
        Double-submit token for cookie-authed state-changing requests.
        Value mirrors the burrow_csrf cookie.
  schemas:
...
    AIGatewayKey:
      type: object
      description: >
        A gateway key as every response but its creation shows it. The key
        itself is stored as a hash and is never returned again.
      properties:
        id: { type: string }
        name: { type: string, maxLength: 120 }
        key_prefix:
          type: string
          description: The first 8 characters of the key, for display.
          example: bgw_Ab3x
        user_id: { type: string, description: The owner. }
        allowed_models:
          type: array
          maxItems: 64
          description: >
            The models the key may use; empty means every model. An entry is
            a synthetic model name, "<provider>/<model>" or "<provider>/*".
          items: { type: string, maxLength: 264 }
        last_used: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
        revoked_at: { type: string, format: date-time, nullable: true }
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message.
...
  responses:
    Unauthorized:
      description: Missing / invalid session or bearer token.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  parameters:
    PathID:
      name: id
      in: path
      required: true
      schema: { type: string }
...
security:
  - CookieAuth: []
    CSRFToken: []
  - BearerAuth: []
paths:
...
  /api/v1/ai/keys:
    get:
      tags: [models]
      summary: List gateway keys
      description: |
        The caller's gateway keys, newest first; an admin sees every user's.
        Revoked keys are listed with revoked_at set. No entry carries the
        key.
      security:
        - CookieAuth: []
        - BearerAuth: []
      responses:
        '200':
          description: The keys.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AIGatewayKey' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [models]
      summary: Create a gateway key
      description: |
        Creates a key owned by the caller. The response carries the key
        ("bgw_" and 43 characters) and is the only one that ever does: the
        relay stores a hash. It is sent with Cache-Control: no-store. The
        body is at most 8 KiB and has no other fields than the two below.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 120
                  description: No control characters.
                allowed_models:
                  type: array
                  maxItems: 64
                  description: >
                    Left out or empty: every model. An entry is a synthetic
                    model name, "<provider>/<model>" or "<provider>/*"; it
                    need not name something that exists.
                  items: { type: string, maxLength: 264 }
            example:
              name: laptop
              allowed_models: [burrow-simple, 'zai/*']
      responses:
        '201':
          description: Key created.
          headers:
            Cache-Control:
              schema: { type: string, enum: [no-store] }
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/AIGatewayKey'
                  - type: object
                    properties:
                      key:
                        type: string
                        description: The key. Shown this once.
        '400':
          description: The body is not valid; the message says why.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/ai/keys/{id}:
    delete:
      tags: [models]
      summary: Revoke a gateway key
      description: |
        The key stops working at once. Its owner or an admin may revoke it;
        another user's key answers 404, like an id that does not exist.
        Revoking a revoked key is a 204 too.
      parameters: [ { $ref: '#/components/parameters/PathID' } ]
      responses:
        '204': { description: Key revoked. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ────────────────────────────────────────────────────────────────────────
  # Database backend status (v0.5.0 Task 15)
  # ────────────────────────────────────────────────────────────────────────
...